Add suport for Windows x64.
Includes support for Windows x64 (AMD64) only, not for IA32 or Mac/Linux x64.

Also includes prject files for Visual Studio 2008, and these project files are the only ones that support the x64 builds.

+++ core/UnicodeUtils.cpp
@@ -304,10 +304,10 @@ namespace avmplus
 			while (c <= 0x7f)
 			{
 				if (--inLen == 0)
-					return outLen+(1+in-inbase);
+					return outLen+(1+(int)(in-inbase));
 				c = *(++in);
 			}
-			outLen += in-inbase;
+			outLen += (int)(in-inbase);
 
 			switch (c >> 4)
 			{
